refactor: 프로젝트 구조 리팩터링 및 ChromaDB 초기화 오류/진단 루프 해결
