+++ src/pipeline/orchestrator.py
@@ -1,443 +1,20 @@
 import json
 import logging
-import pickle
 import threading
-import time
-import uuid
-from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any
 
 from src.common.config import settings
-from src.common.constants import MetadataFields
 from src.core.cache import SemanticCache
-from src.data.parser import ManualParser
-from src.processing.chunking import HierarchicalChunker, create_parent_child_chunks
-from src.processing.pdf_parser import DoclingPDFParser
+from src.core.storage import StorageManager
+from src.pipeline.ingestion import IngestionPipeline
+from src.pipeline.strategies import DoclingPDFParserStrategy, ManualParserStrategy, ParserStrategy
 from src.utils.file_utils import generate_file_hash
-from src.utils.logger import TracingLogger, setup_global_logging
-from src.utils.paths import CACHE_DIR, PROCESSED_DATA_DIR, RAW_DATA_DIR, ensure_directories
-from src.vector_db.chroma_manager import ChromaDBManager
+from src.utils.logger import TracingLogger
+from src.utils.paths import CACHE_DIR, PROCESSED_DATA_DIR, RAW_DATA_DIR
+from src.utils.unicode import normalize_path_to_nfc, normalize_to_nfc
 
-# 로깅 설정
 logger = logging.getLogger(__name__)
-
-
-class ParserStrategy(ABC):
-    """문서 파싱 전략을 위한 추상 베이스 클래스"""
-
-    @abstractmethod
-    def parse(self, file_path: Path) -> list[dict[str, Any]]:
-        pass
-
-
-class ManualParserStrategy(ParserStrategy):
-    """기존 ManualParser를 사용하는 전략"""
-
-    def parse(self, file_path: Path) -> list[dict[str, Any]]:
-        # ManualParser는 RAW_DATA_DIR 기준 상대 경로를 받음
-        relative_path = file_path.relative_to(RAW_DATA_DIR)
-        parser = ManualParser(str(relative_path), parser_type="manual")
-        return parser.parse()
-
-
-class MarkdownParserStrategy(ParserStrategy):
-    """Markdown 파일을 구조 파괴 없이 읽어오는 전략"""
-
-    def parse(self, file_path: Path) -> list[dict[str, Any]]:
-        with open(file_path, encoding="utf-8") as f:
-            raw_md_text = f.read()
-
-        if len(raw_md_text.strip()) < 5:
-            logger.warning(f"마크다운 파일의 내용이 너무 짧아 건너뜁니다: {file_path.name}")
-            return []
-
-        parser_type = settings.PARSER_TYPE.lower()
-        base_metadata = {
-            MetadataFields.SOURCE_ID: generate_file_hash(file_path, parser_type),
-            MetadataFields.SRC_NAME: file_path.name,
-            MetadataFields.RELATIVE_PATH: str(file_path.relative_to(RAW_DATA_DIR)),
-            MetadataFields.PARSER_TYPE: parser_type,
-            MetadataFields.DOC_TYPE: file_path.suffix.lower().replace(".", ""),
-            MetadataFields.PG_NUM: 1,
-            MetadataFields.CATEGORY: file_path.parent.name if file_path.parent.name != "raw" else "일반",
-        }
-
-        return [{"is_raw_markdown": True, "content": raw_md_text, "metadata": base_metadata}]
-
-
-class DoclingPDFParserStrategy(ParserStrategy):
-    """
-    IBM Docling 기반 고품질 PDF 파서 전략.
-    - 표 구조 34개+ 자동 감지 및 마크다운 변환
-    - 한글 OCR (RapidOCR) 지원
-    - 레이아웃 구조 보존 (단일 Converter 인스턴스 재사용으로 메모리 효율 확보)
-    """
-
-    def __init__(self):
-        self.pdf_parser = DoclingPDFParser()  # AI 모델 1회만 로드
-        self.manual_parser = ManualParser  # PDF 외 파일 (MD 등) Fallback
-
-    def parse(self, file_path: Path) -> list[dict[str, Any]]:
-        if file_path.suffix.lower() == ".pdf":
-            # 1. 캐시 파일 경로 설정
-            source_id = generate_file_hash(file_path, parser_type="docling")
-            cache_file = CACHE_DIR / f"{source_id}_parsed.pkl"
-
-            # 2. 캐시 존재 시 무거운 파싱 생략
-            if cache_file.exists():
-                logger.info(f"캐시된 Docling 파싱 결과를 로드합니다: {file_path.name}")
-                with open(cache_file, "rb") as f:
-                    return pickle.load(f)
-
-            logger.info(f"DoclingPDFParser를 사용하여 PDF 파싱: {file_path.name}")
-            start_time = time.time()
-            parsed = self.pdf_parser.parse(file_path)  # {markdown, tables, page_count, table_count}
-            elapsed = time.time() - start_time
-            logger.info(f"파싱 완료: {file_path.name} (소요 시간: {elapsed:.2f}초, 표 {parsed['table_count']}개 감지)")
-
-            # 3. Docling 결과를 pipeline 호환 포맷으로 매핑
-            #    is_combined=True 플래그로 create_parent_child_chunks 경로를 타도록 지정
-            results = [
-                {
-                    "is_combined": True,
-                    "content": parsed["markdown"],
-                    "metadata": {
-                        MetadataFields.SOURCE_ID: source_id,
-                        MetadataFields.SRC_NAME: file_path.name,
-                        MetadataFields.RELATIVE_PATH: str(file_path.relative_to(RAW_DATA_DIR)),
-                        MetadataFields.PARSER_TYPE: "docling",
-                        MetadataFields.PG_NUM: 1,
-                        MetadataFields.DOC_TYPE: "pdf",
-                        MetadataFields.CATEGORY: file_path.parent.name,
-                        # Docling 전용 메타데이터: 표 감지 정보
-                        "table_count": parsed["table_count"],
-                        "page_count": parsed["page_count"],
-                        "has_table": parsed["table_count"] > 0,
-                    },
-                }
-            ]
-
-            # 4. 캐시 저장
-            with open(cache_file, "wb") as f:
-                pickle.dump(results, f)
-
-            return results
-        else:
-            # PDF가 아닌 경우 ManualParser로 Fallback
-            relative_path = file_path.relative_to(RAW_DATA_DIR)
-            parser = self.manual_parser(str(relative_path), parser_type="docling")
-            return parser.parse()
-
-
-class IngestionPipeline:
-    """데이터 전처리 및 인덱싱을 담당하는 클래스"""
-
-    def __init__(
-        self,
-        strategy: ParserStrategy,
-        raw_dir: Path = RAW_DATA_DIR,
-        processed_dir: Path = PROCESSED_DATA_DIR,
-        collection_name: str = "rag_collection",
-    ):
-        self.strategy = strategy
-        self.raw_dir = raw_dir
-        self.processed_dir = processed_dir
-        self.chunker = HierarchicalChunker()
-        # ChromaDBManager 인스턴스를 초기화 시 한 번만 생성하여 재사용
-        self.db_manager = ChromaDBManager(collection_name=collection_name)
-        ensure_directories()
-
-    def scan_files(self, supported_exts: list[str] | None = None) -> list[Path]:
-        if supported_exts is None:
-            supported_exts = [".pdf", ".md", ".markdown"]
-        files = []
-        for ext in supported_exts:
-            # 모든 하위 디렉토리를 포함하여 검색
-            all_files = list(self.raw_dir.glob(f"**/*{ext}"))
-            # Mac 숨김 파일(._ 로 시작) 및 시스템 파일 필터링
-            filtered_files = [f for f in all_files if not f.name.startswith("._") and f.name != ".DS_Store"]
-            files.extend(filtered_files)
-        return files
-
-    def process_and_chunk(  # noqa: C901
-        self, files: list[Path], progress_callback=None, file_parser_types: dict[str, str] | None = None
-    ) -> list[dict[str, Any]]:
-        all_hierarchical_data = []
-        total_files = len(files)
-        for idx, file_path in enumerate(files):
-            if progress_callback:
-                try:
-                    progress_callback(idx, total_files, file_path.name)
-                except Exception as cb_e:
-                    logger.error(f"진행 상황 콜백 호출 실패: {cb_e}")
-            try:
-                # 1. 파일 확장자에 따른 전략 동적 선택 및 파싱
-                if file_path.suffix.lower() == ".pdf":
-                    import unicodedata
-
-                    rel_path = unicodedata.normalize("NFC", str(file_path.relative_to(RAW_DATA_DIR)))
-                    normalized_parser_types = {
-                        unicodedata.normalize("NFC", k): v for k, v in (file_parser_types or {}).items()
-                    }
-                    file_parser = normalized_parser_types.get(rel_path, "manual").lower()
-                    logger.warning(
-                        f"=== 디버깅 파서 선택 ===\n"
-                        f"파일명: {file_path.name}\n"
-                        f"rel_path (NFC): {rel_path}\n"
-                        f"결정된 file_parser: {file_parser}\n"
-                        f"전달된 file_parser_types: {file_parser_types}\n"
-                        f"정규화된 parser_types: {normalized_parser_types}\n"
-                        f"========================="
-                    )
-                    if file_parser == "docling":
-                        try:
-                            import docling  # noqa: F401
-
-                            active_strategy = DoclingPDFParserStrategy()
-                        except ImportError:
-                            logger.error("docling 라이브러리가 없어 manual 전략으로 대체합니다.")
-                            active_strategy = ManualParserStrategy()
-                    else:
-                        active_strategy = ManualParserStrategy()
-                else:
-                    active_strategy = MarkdownParserStrategy()
-
-                sections = active_strategy.parse(file_path)
-
-                if not sections:
-                    continue
-
-                # 2. 계층적 청킹 (결과 타입별 분기 처리)
-                if sections[0].get("is_raw_markdown"):
-                    file_chunks = create_parent_child_chunks(sections[0]["content"], sections[0]["metadata"])
-                    all_hierarchical_data.extend(file_chunks)
-                elif sections[0].get("is_combined"):
-                    for sec in sections:
-                        file_chunks = create_parent_child_chunks(sec["content"], sec["metadata"])
-                        all_hierarchical_data.extend(file_chunks)
-                else:
-                    # 기존 ManualParser PDF 처리 방식 (Fallback 호환성)
-                    for sec in sections:
-                        parent_id = str(uuid.uuid4())
-                        meta_for_children = sec["metadata"].copy()
-                        sec_title = f"{sec.get('chapter', '기본 섹션')} > {sec.get('article', '기본 섹션')}"
-                        meta_for_children[MetadataFields.SEC_TITLE] = sec_title
-                        meta_for_children[MetadataFields.HEADER_PATH] = sec_title
-
-                        children = self.chunker.split_into_children(sec["content"], parent_id, meta_for_children)
-
-                        if children:
-                            parent_metadata = {
-                                MetadataFields.SOURCE_ID: sec["metadata"].get(MetadataFields.SOURCE_ID, "UNKNOWN"),
-                                MetadataFields.SRC_NAME: sec["metadata"].get(MetadataFields.SRC_NAME, "UNKNOWN"),
-                                MetadataFields.DOC_TYPE: sec["metadata"].get(MetadataFields.DOC_TYPE, "pdf"),
-                                MetadataFields.PG_NUM: sec["metadata"].get(MetadataFields.PG_NUM, 1),
-                                MetadataFields.SEC_TITLE: sec_title,
-                                MetadataFields.CHUNK_ID: parent_id,
-                                MetadataFields.PARENT_ID: None,
-                                MetadataFields.HEADER_PATH: sec_title,
-                                MetadataFields.IS_TABLE: False,
-                            }
-                            all_hierarchical_data.append(
-                                {
-                                    "parent_id": parent_id,
-                                    "parent_text": sec["content"],
-                                    "metadata": parent_metadata,
-                                    "children": children,
-                                }
-                            )
-
-            except Exception as e:
-                logger.error(f"파일 처리 실패: {file_path.name} - {e!s}")
-
-        if progress_callback and total_files > 0:
-            try:
-                progress_callback(total_files, total_files, "모든 파일 처리 완료")
-            except Exception as cb_e:
-                logger.error(f"진행 상황 콜백 호출 실패: {cb_e}")
-
-        return all_hierarchical_data
-
-    def save_processed_data(self, data: list[dict[str, Any]]) -> list[Path]:
-        """
-        전처리된 데이터를 source_id별 개별 JSON 파일로 저장합니다.
-        (BM25 인덱스와 정합성 유지를 위해 개별 파일 관리가 필수적임)
-        """
-        saved_paths = []
-        # 데이터를 source_id별로 그룹화
-        grouped_data = {}
-        for item in data:
-            sid = item["metadata"].get(MetadataFields.SOURCE_ID, "unknown")
-            if sid not in grouped_data:
-                grouped_data[sid] = []
-            grouped_data[sid].append(item)
-
-        for sid, sid_data in grouped_data.items():
-            save_path = self.processed_dir / f"{sid}.json"
-            with open(save_path, "w", encoding="utf-8") as f:
-                json.dump(sid_data, f, ensure_ascii=False, indent=2)
-            saved_paths.append(save_path)
-            logger.info(f"   - 전처리 결과 저장: {save_path.name}")
-
-        return saved_paths
-
-    def _prepare_cleanup_targets(
-        self, source_ids_to_delete: list[str] | None, filenames_to_delete: list[str] | None
-    ) -> tuple[list[str], list[str]]:
-        import unicodedata
-        from pathlib import Path
-
-        valid_ids = list(set([sid for sid in (source_ids_to_delete or []) if sid]))
-
-        target_filenames = []
-        if filenames_to_delete:
-            for fname in filenames_to_delete:
-                if not fname:
-                    continue
-                name_only = Path(fname).name
-                nfc_n = unicodedata.normalize("NFC", name_only)
-                nfd_n = unicodedata.normalize("NFD", name_only)
-                target_filenames.extend([nfc_n, nfd_n])
-
-        target_filenames = list(set(target_filenames))
-
-        if target_filenames:
-            logger.info(f"파일명 기반 클린업 대상 확인: {filenames_to_delete}")
-            try:
-                results = self.db_manager.collection.get(
-                    where={"src_name": {"$in": target_filenames}}, include=["metadatas"]
-                )
-                if results and results.get("metadatas"):
-                    for meta in results["metadatas"]:
-                        if meta and "source_id" in meta:
-                            valid_ids.append(meta["source_id"])
-            except Exception as e:
-                logger.error(f"파일명 기반 source_id 조회 중 오류 발생: {e}")
-
-        return list(set(valid_ids)), target_filenames
-
-    def cleanup_db(
-        self,
-        source_ids_to_delete: list[str] | None = None,
-        relative_paths_to_delete: list[str] | None = None,
-        filenames_to_delete: list[str] | None = None,
-    ):
-        """
-        DB, 캐시 및 가공된 JSON 파일에서 삭제된 데이터를 제거합니다.
-        (BM25Manager는 processed_dir의 모든 json을 읽으므로 물리적 파일 삭제가 곧 정합성임)
-        """
-        valid_ids, target_filenames = self._prepare_cleanup_targets(source_ids_to_delete, filenames_to_delete)
-
-        if not valid_ids and not relative_paths_to_delete and not target_filenames:
-            return
-
-        logger.info("데이터 정합성 검증 및 클린업을 수행합니다.")
-
-        # 1. 벡터 DB 데이터 삭제
-        if target_filenames:
-            try:
-                self.db_manager.delete_documents(where={"src_name": {"$in": target_filenames}})
-                logger.info(f"   - ChromaDB 파일명 기준 {len(target_filenames)}개 파일 삭제 완료")
-            except Exception as e:
-                logger.error(f"ChromaDB 파일명 기준 삭제 실패: {e}")
-
-        self._delete_from_vector_db(valid_ids, relative_paths_to_delete)
-
-        # 2. 물리적 캐시 및 JSON 파일 삭제
-        deleted_count = self._delete_physical_files(valid_ids, relative_paths_to_delete)
-
-        if deleted_count > 0:
-            logger.info(f"물리적 데이터 파일 {deleted_count}개 삭제 완료")
-
-        logger.info("데이터 정합성 검증 및 클린업 작업이 완료되었습니다.")
-
-    def _delete_from_vector_db(self, source_ids_to_delete: list[str], relative_paths_to_delete: list[str] | None):
-        """벡터 DB에서 데이터 삭제"""
-        # 상대 경로 기반 삭제 (원자적 정리 강화)
-        if relative_paths_to_delete:
-            for rel_path in relative_paths_to_delete:
-                logger.info(f"   - 상대 경로 기준 삭제: {rel_path}")
-                self.db_manager.delete_documents(where={MetadataFields.RELATIVE_PATH: rel_path})
-
-        # Source ID 기반 삭제 (하위 호환성)
-        if source_ids_to_delete:
-            self.db_manager.delete_documents(where={MetadataFields.SOURCE_ID: {"$in": source_ids_to_delete}})
-            logger.info(f"   - ChromaDB {len(source_ids_to_delete)}개 Source ID 삭제 완료")
-
-    def _delete_physical_files(
-        self, source_ids_to_delete: list[str], relative_paths_to_delete: list[str] | None
-    ) -> int:
-        """물리적 캐시 및 JSON 파일 삭제"""
-        deleted_count = 0
-
-        # a. Source ID 기반 삭제
-        for sid in source_ids_to_delete:
-            if self._remove_sid_files(sid):
-                deleted_count += 1
-
-        # b. 상대 경로 기반 물리 파일 삭제 (파서 변경 등 중복 방지)
-        if relative_paths_to_delete:
-            deleted_count += self._remove_files_by_rel_path(relative_paths_to_delete)
-
-        return deleted_count
-
-    def _remove_sid_files(self, sid: str) -> bool:
-        """특정 Source ID와 관련된 물리 파일 삭제"""
-        removed = False
-        # 파싱 캐시 삭제
-        cache_file = CACHE_DIR / f"{sid}_parsed.pkl"
-        if cache_file.exists():
-            cache_file.unlink()
-
-        # 가공된 JSON 삭제 (BM25 정합성 핵심)
-        json_file = self.processed_dir / f"{sid}.json"
-        if json_file.exists():
-            json_file.unlink()
-            removed = True
-        return removed
-
-    def _remove_files_by_rel_path(self, relative_paths_to_delete: list[str]) -> int:
-        """상대 경로 목록에 해당하는 물리 파일들을 찾아 삭제"""
-        deleted_count = 0
-        for json_file in list(self.processed_dir.glob("*.json")):
-            if json_file.name == "manifest.json":
-                continue
-            try:
-                data = None
-                with open(json_file, encoding="utf-8") as f:
-                    data = json.load(f)
-
-                if not (data and isinstance(data, list) and len(data) > 0):
-                    continue
-
-                meta = data[0].get("metadata", {})
-                if meta.get(MetadataFields.RELATIVE_PATH) in relative_paths_to_delete:
-                    sid = json_file.stem
-                    # 캐시 삭제
-                    cache_file = CACHE_DIR / f"{sid}_parsed.pkl"
-                    if cache_file.exists():
-                        cache_file.unlink()
-                    # JSON 삭제
-                    if json_file.exists():
-                        json_file.unlink()
-                        deleted_count += 1
-            except Exception as e:
-                logger.warning(f"파일 스캔 중 오류 ({json_file.name}): {e}")
-        return deleted_count
-
-    def upsert_to_db(self, data: list[dict[str, Any]]):
-        ids, docs, metas = [], [], []
-        for parent in data:
-            for child in parent["children"]:
-                ids.append(child["chunk_id"])
-                docs.append(child["text"])
-                metas.append(child["metadata"])
-
-        if ids:
-            logger.info(f"ChromaDB 업서트 시작 ({len(ids)}개 청크)...")
-            self.db_manager.upsert_documents(ids=ids, documents=docs, metadatas=metas)
-            logger.info("ChromaDB 업서트 완료!")
 
 
 class PipelineOrchestrator:
@@ -463,7 +40,9 @@ class PipelineOrchestrator:
             logger.info("PipelineOrchestrator 초기화...")
             self.parser_type = settings.PARSER_TYPE.lower()
             self.strategy = self._get_parser_strategy()
-            self.ingestion_pipeline = IngestionPipeline(self.strategy)
+            # StorageManager 인스턴스 생성 및 IngestionPipeline에 전달
+            self.storage_manager = StorageManager(PROCESSED_DATA_DIR, CACHE_DIR)
+            self.ingestion_pipeline = IngestionPipeline(self.strategy, storage_manager=self.storage_manager)
             self.tracing_logger = TracingLogger()
             self.manifest_path = PROCESSED_DATA_DIR / "manifest.json"
             self.cache = SemanticCache()
@@ -530,14 +109,12 @@ class PipelineOrchestrator:
 
         new_files = {}
         for f in all_files:
-            import unicodedata
-
-            rel_path = unicodedata.normalize("NFC", str(f.relative_to(RAW_DATA_DIR)))
+            rel_path = normalize_path_to_nfc(f.relative_to(RAW_DATA_DIR))
             old_info = old_files.get(rel_path, {})
             if not old_info:
                 # NFD-NFC 매치 백업
                 for k, v in old_files.items():
-                    if unicodedata.normalize("NFC", k) == rel_path:
+                    if normalize_to_nfc(k) == rel_path:
                         old_info = v
                         break
             file_parser_type = old_info.get("parser_type", self.parser_type)
@@ -708,15 +285,16 @@ class PipelineOrchestrator:
                     logger.info("ChromaDB 컬렉션 및 가공 파일 물리적 초기화 시작...")
                     self.ingestion_pipeline.db_manager.reset_collection()
 
-                    for f in self.ingestion_pipeline.processed_dir.glob("*.json"):
+                    # processed_dir의 JSON 파일들 및 CACHE_DIR의 pkl 파일들 정리
+                    for f in self.storage_manager.scan_processed_files():
                         try:
-                            f.unlink()
+                            self.storage_manager.delete_file(f)
                         except Exception as e:
                             logger.error(f"JSON 파일 삭제 실패 {f}: {e}")
 
                     for f in CACHE_DIR.glob("*_parsed.pkl"):
                         try:
-                            f.unlink()
+                            self.storage_manager.delete_file(f)
                         except Exception as e:
                             logger.error(f"캐시 파일 삭제 실패 {f}: {e}")
 
@@ -725,10 +303,8 @@ class PipelineOrchestrator:
                     relative_paths_to_delete = None
 
                     new_files = {}
-                    import unicodedata
-
                     for f in all_files:
-                        rel_path = unicodedata.normalize("NFC", str(f.relative_to(RAW_DATA_DIR)))
+                        rel_path = normalize_path_to_nfc(f.relative_to(RAW_DATA_DIR))
                         new_files[rel_path] = {
                             "hash": generate_file_hash(f, self.parser_type),
                             "parser_type": self.parser_type,
@@ -815,12 +391,10 @@ class PipelineOrchestrator:
         old_manifest = self._load_manifest()
         old_files = old_manifest.get("files", {})
 
-        import unicodedata
-
-        normalized_file_name = unicodedata.normalize("NFC", file_name)
+        normalized_file_name = normalize_to_nfc(file_name)
         target_rel_path = None
         for rel_path in old_files:
-            normalized_rel_path = unicodedata.normalize("NFC", rel_path)
+            normalized_rel_path = normalize_to_nfc(rel_path)
             is_match = (
                 Path(normalized_rel_path).name == normalized_file_name or normalized_rel_path == normalized_file_name
             )
@@ -830,9 +404,9 @@ class PipelineOrchestrator:
 
         if not target_rel_path:
             for f in self.ingestion_pipeline.scan_files():
-                normalized_f_name = unicodedata.normalize("NFC", f.name)
+                normalized_f_name = normalize_to_nfc(f.name)
                 if normalized_f_name == normalized_file_name:
-                    target_rel_path = unicodedata.normalize("NFC", str(f.relative_to(RAW_DATA_DIR)))
+                    target_rel_path = normalize_path_to_nfc(f.relative_to(RAW_DATA_DIR))
                     break
 
         if not target_rel_path:
@@ -866,13 +440,12 @@ class PipelineOrchestrator:
         hashes_to_delete = []
         for file_name, new_parser in file_parser_map.items():
             new_parser_lower = new_parser.lower()
-            import unicodedata
 
-            normalized_file_name = unicodedata.normalize("NFC", file_name)
+            normalized_file_name = normalize_to_nfc(file_name)
 
             target_rel_path = None
             for rel_path in old_files:
-                normalized_rel_path = unicodedata.normalize("NFC", rel_path)
+                normalized_rel_path = normalize_to_nfc(rel_path)
                 is_match = (
                     Path(normalized_rel_path).name == normalized_file_name
                     or normalized_rel_path == normalized_file_name
@@ -883,9 +456,9 @@ class PipelineOrchestrator:
 
             if not target_rel_path:
                 for f in self.ingestion_pipeline.scan_files():
-                    normalized_f_name = unicodedata.normalize("NFC", f.name)
+                    normalized_f_name = normalize_to_nfc(f.name)
                     if normalized_f_name == normalized_file_name:
-                        target_rel_path = unicodedata.normalize("NFC", str(f.relative_to(RAW_DATA_DIR)))
+                        target_rel_path = normalize_path_to_nfc(f.relative_to(RAW_DATA_DIR))
                         break
 
             if target_rel_path:
@@ -927,18 +500,3 @@ class PipelineOrchestrator:
                     target_paths.append(RAW_DATA_DIR / target_rel_path)
 
             self.run_ingestion(force=False, progress_callback=progress_callback, target_files=target_paths)
-
-
-# 하위 호환성을 위한 기존 클래스 래핑
-class PreprocessingPipeline:
-    def __init__(self, *args, **kwargs):
-        self.orchestrator = PipelineOrchestrator()
-
-    def run(self, *args, **kwargs):
-        return self.orchestrator.run_ingestion()
-
-
-if __name__ == "__main__":
-    setup_global_logging()
-    orchestrator = PipelineOrchestrator()
-    orchestrator.run_ingestion()
